+++ custom_components/intervals_icu/const.py
@@ -15,3 +15,22 @@ DEFAULT_SCAN_INTERVAL_MINUTES = 30
 MIN_SCAN_INTERVAL_MINUTES = 5
 
 DATA_COORDINATOR = "coordinator"
+
+SERVICE_SET_WELLNESS = "set_wellness"
+
+ATTR_DATE = "date"
+ATTR_WEIGHT = "weight"
+ATTR_KCAL_CONSUMED = "kcal_consumed"
+ATTR_CARBOHYDRATES = "carbohydrates"
+ATTR_PROTEIN = "protein"
+ATTR_FAT_TOTAL = "fat_total"
+ATTR_HYDRATION_VOLUME = "hydration_volume"
+
+WRITABLE_WELLNESS_FIELD_MAP: dict[str, str] = {
+    ATTR_WEIGHT: "weight",
+    ATTR_KCAL_CONSUMED: "kcalConsumed",
+    ATTR_CARBOHYDRATES: "carbohydrates",
+    ATTR_PROTEIN: "protein",
+    ATTR_FAT_TOTAL: "fatTotal",
+    ATTR_HYDRATION_VOLUME: "hydrationVolume",
+}